Move development subdomain logic into ENV rather than headers
Previously we took a naive approach to overriding subdomains in local
environments by patching the HOST (and tentatively ORIGIN) headers. But
this caused problems with a) CSRF authenticity tokens and b) follow-up
HTTP redirects.

Instead, we now just assume that the channel subdomain can be overriden
in any environment by setting the `env["marathonstream.subdomain"]`
value and then only including the middleware that sets that value in the
development environment