Introduce authorization
We don't yet have a way to set up a user, but we have an API for the
user object to conform to when it's been created.

And we're abstracting the actual user object away from the authorized
actions so when we implement the authentication it shouldn't have any
effect on the actions (fingers-crossed-emoji)

To do this we're making a couple of shortcuts to allow these actions to
be tested. These shortcuts are pretty isolated but until we have an
actual OAuth endpoint to simulate, they're technically a little hacky