Correctly recognise current_session in child controllers
When we extracted `current_session` into the Channelable concern, this
broke the `pundit_user` method. Because we implemented it as an `alias`,
it was created as a copy of the "super" method, and ignored the
"subclassed" method with the same name. By rewriting this as a regular
method call, the name is resolved at runtime which means it can defer to
the method in the Channelable module.