Add support for exhaustive role testing
This commit adds hooks into the Minitest spec library so that exhaustive
test blocks can be run.

The way this works is that inside a `describe` block, the test helpers
keep track of which roles have been tested using `as`. Then at the end
of the block, if any roles are missing then a new always-failing test is
inserted with details.

The whole thing is sneaky metaprogramming and probably not a good idea.