Switch to dynamic subdomain parameter
Previously, hardcoding the subdomain parameter into links as _sudo was
fine for development, but would have broken when the app was deployed.

Now, we introduce the parameter name as a config variable.

This works because Rails lets you pass the subdomain in the same way as
it lets you pass any other parameter.