Include cgi gem to override Ruby stdlib version
Ruby stdlib doesn't allow cookies to be set with domains with leading
dots. Rails, unfortunately, sets cookies with domains with leading dots
(this is technically incorrect to do)

A newer version of the CGI library that handles this case is available,
but only by including the gem version of the library.

See: https://github.com/ruby/cgi/pull/29