Improve test screenshot filenames
Previously they didn't include the test class, just the test name, which
led to ambiguity in screenshot names.

In particular, each `as a` block makes a new test class, which means any
test screenshot from one of these blocks didn't leave any way to show
which test it was from.