Refactor test authentication helpers
Most of the authentication test helpers we wrote already were generic,
but in particular the login steps work differently between integration
(i.e. controller) tests and system (capybara/selenium) tests.

This commit extracts the common parts of the functionality and
introduces specific minimal modules for each of the test types.